+++ tests/s3handler_tests/delete_test.go
